refactor(where): Areas and Places are one surface
Three design reviews, run independently, all reached the same objection to
the nav item I added yesterday: Places is a thin, set-and-forget list that
bought permanent nav space, and the pin and the named places were still on
two different pages. The cross-link I added from the Areas page was the tell
— it existed because the content was in the wrong place.

- The Places page becomes a section of the Areas page, directly under the
  Location card that holds the pin. It no longer renders a pin card of its
  own; the card above it is the pin.
- One nav item, relabelled Areas & Places, replacing two. The page title
  matches. /places redirects rather than 404s.
- The notification language block moved from the middle of the page to the
  end. It has nothing to do with areas and was interrupting the picker.

Two of the three reviews argued AGAINST merging Areas and My Geofences, and
the third named that as the strongest counter to its own proposal, so
Geofences stays its own item. It hosts drawing, submit-for-review and
approval states: a workspace, not a list.

Left undone deliberately: Areas and My Geofences both edit profiles.area
through different controls, chips on one page and slide toggles on the other.
Verified rather than taken on trust. That is a deeper incoherence than
anything this change touches and wants its own decision.
