+++ Applications/Pgan.PoracleWebNet.App/ClientApp/src/app/modules/areas/area-list.component.html
@@ -63,10 +63,6 @@
               <mat-icon>edit_location</mat-icon>
               {{ location() ? ('AREAS.CHANGE_LOCATION' | translate) : ('AREAS.SET_LOCATION' | translate) }}
             </button>
-            <a mat-button routerLink="/places">
-              <mat-icon>place</mat-icon>
-              {{ 'AREAS.MANAGE_PLACES' | translate }}
-            </a>
           </div>
         </div>
         <p class="method-description">
@@ -84,17 +80,7 @@
       <span>{{ 'AREAS.METHOD_NOTE' | translate }}</span>
     </div>
 
-    <!-- Notification Language -->
-    <div class="notification-language">
-      <div class="notification-language-text">
-        <h3>
-          <mat-icon>translate</mat-icon>
-          {{ 'AREAS.NOTIFICATION_LANGUAGE' | translate }}
-        </h3>
-        <p>{{ 'AREAS.NOTIFICATION_LANGUAGE_DESC' | translate }}</p>
-      </div>
-      <app-language-selector></app-language-selector>
-    </div>
+    <app-places-section />
 
     <!-- Area Selection Section -->
     <h2 class="section-title">
@@ -231,5 +217,17 @@
         </button>
       </div>
     }
+
+    <!-- Notification Language -->
+    <div class="notification-language">
+      <div class="notification-language-text">
+        <h3>
+          <mat-icon>translate</mat-icon>
+          {{ 'AREAS.NOTIFICATION_LANGUAGE' | translate }}
+        </h3>
+        <p>{{ 'AREAS.NOTIFICATION_LANGUAGE_DESC' | translate }}</p>
+      </div>
+      <app-language-selector></app-language-selector>
+    </div>
   </div>
 }
